Exctract paser methods
